+++ buttplug/src/server/device/protocol/kiiroo_v21.rs
@@ -93,7 +93,7 @@ impl ProtocolHandler for KiirooV21 {
     let previous_position = self.previous_position.load(Relaxed);
     let distance = (previous_position as f64 - (position as f64)).abs() / 99f64;
     let position = position as u8;
-    let speed = (calculate_speed(distance, duration as u32) * 99f64) as u8;
+    let speed = (calculate_speed(distance, duration) * 99f64) as u8;
     self.previous_position.store(position, Relaxed);
     Ok(vec![HardwareWriteCmd::new(
       feature_id,
